Rename unsignedPayload and ec2QueryName
Now that the aws.api#protocols and aws.api#auth namespaces exist, these
traits make more sense to be defined in those namespaces.
