+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Smithy Changelog
 
+## 0.10.0 (2020-??-??)
+
+### Breaking changes
+
+* aws.api#ec2QueryName was renamed to `aws.protocols#ec2QueryName`.
+* aws.api#unsignedPayload was renamed to `aws.auth#unsignedPayload` and no longer takes an array of values.
+  All auth schemes that honor this trait should be considered unsigned when this trait is applied.
+
+
 ## 0.9.7 (2020-01-15)
 
 ### Features
