+++ docs/source/spec/aws-query-protocol.rst
@@ -102,4 +102,66 @@ Value type
             }
         }
 
+
+.. _aws.protocols#ec2QueryName-trait:
+
+------------------------------------
+``aws.protocols#ec2QueryName`` trait
+------------------------------------
+
+Summary
+    Indicates the serialized name of a structure member when that structure is
+    serialized for the input of an EC2 operation using the
+    ``aws.protocols#ec2Query`` protocol.
+Trait selector
+    ``member:of(structure)``
+Value type
+    ``string``
+
+It is very important to note that the ``aws.protocols#ec2QueryName`` ONLY applies
+when serializing an INPUT. For example, given the following Smithy model:
+
+.. tabs::
+
+    .. code-tab:: smithy
+
+        structure MyStruct {
+            @ec2QueryName("foo")
+            bar: String
+        }
+
+    .. code-tab:: json
+
+        {
+            "smithy": "0.5.0",
+            "shapes": {
+                "smithy.example#MyStruct": {
+                    "type": "structure",
+                    "members": {
+                        "bar": {
+                            "target": "smithy.api#String",
+                            "traits": {
+                                "aws.protocols#ec2QueryName": "foo"
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+The serialization of this structure as an input is:
+
+::
+
+    MyStruct.bar=baz
+
+The serialization of the structure as an (XML) output is:
+
+.. code-block:: xml
+
+    <MyStruct>
+        <foo>baz</foo>
+    </MyStruct>
+
+
 *TODO: Add specifications, protocol examples, etc.*
