+++ smithy-aws-protocol-tests/model/ec2Query/input.smithy
@@ -4,7 +4,7 @@ $version: "0.5.0"
 
 namespace aws.protocoltests.ec2
 
-use aws.api#ec2QueryName
+use aws.protocols#ec2QueryName
 use aws.protocols#ec2Query
 use aws.protocoltests.shared#EpochSeconds
 use aws.protocoltests.shared#FooEnum
