+++ smithy-aws-traits/src/main/resources/META-INF/smithy/aws.api.json
@@ -73,15 +73,6 @@
                 "smithy.api#documentation": "Marks a string as containing an ARN."
             }
         },
-        "aws.api#unsignedPayload": {
-            "type": "structure",
-            "traits": {
-                "smithy.api#trait": {
-                    "selector": "operation"
-                },
-                "smithy.api#documentation": "Indicates that the request payload of a signed request is not to be used as part of the signature."
-            }
-        },
         "aws.api#data": {
             "type": "string",
             "traits": {
@@ -217,16 +208,6 @@
                     "selector": "service"
                 },
                 "smithy.api#documentation": "Configures endpoint discovery for the service."
-            }
-        },
-        "aws.api#ec2QueryName": {
-            "type": "string",
-            "traits": {
-                "smithy.api#trait": {
-                    "selector": "member:of(structure)"
-                },
-                "smithy.api#documentation": "Indicates the serialized name of a structure member when that structure is serialized for the input of an EC2 operation.",
-                "smithy.api#pattern": "^[a-zA-Z_][a-zA-Z_0-9-]*$"
             }
         }
     }
