+++ smithy-aws-traits/src/test/resources/software/amazon/smithy/aws/traits/auth/unsigned-request-payload.json
@@ -4,7 +4,7 @@
         "ns.foo#Unsigned1": {
             "type": "operation",
             "traits": {
-                "aws.api#unsignedPayload": true
+                "aws.auth#unsignedPayload": true
             }
         }
     }
