+++ smithy-aws-traits/src/test/resources/software/amazon/smithy/aws/traits/errorfiles/ec2-query-name.json
@@ -8,7 +8,7 @@
                     "target": "smithy.api#String",
                     "traits": {
                         "smithy.api#xmlName": "ipv6AddressSet",
-                        "aws.api#ec2QueryName": "Ipv6Addresses"
+                        "aws.protocols#ec2QueryName": "Ipv6Addresses"
                     }
                 }
             }
